mysite/requirements.txt
django==2.0.8
djangorestframework==3.7.3
psycopg2==2.7.3
celery==3.1.18
redis==2.10.6
